third_party/blink/web_tests/http/tests/xsl/xslt/permissions-policy-frame-policy-in-xsl-quirk.html

<body>
  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>
  <script>
    const iframe = document.createElement('iframe');
    iframe.allow = "geolocation 'none'";
    iframe.src = "resources/permissions-policy-iframe.xml";

    async_test(t => {
      window.addEventListener('message', t.step_func_done(event => {
        assert_equals(event.data, 'geolocation disallowed');
      }), {once: true});
    });

    document.body.appendChild(iframe);
  </script>
</body>
